requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
iniconfig==2.0.0
model-bakery==1.20.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
sqlparse==0.5.1
